Series_App/requirements.txt
pandas
pyHIve
matplotlib.pyplot
matplotlib.ticker
textblob
preprocessor
streamlit